move as_toml_string() functionality into bapsf_motion.utils.toml
