+++ bapsf_motion/actors/manager_.py
@@ -254,20 +254,7 @@ class RunManagerConfig(UserDict):
 
     @property
     def as_toml_string(self) -> str:
-        def convert_key_to_string(_d):
-            _config = {}
-            for key, value in _d.items():
-                if isinstance(value, (dict, UserDict)):
-                    value = convert_key_to_string(value)
-
-                if not isinstance(key, str):
-                    key = f"{key}"
-
-                _config[key] = value
-
-            return _config
-
-        return "[run]\n" + toml.dumps(convert_key_to_string(self))
+        return "[run]\n" + toml.as_toml_string(self)
 
     def update_run_name(self, name: str):
         if not isinstance(name, str):
